fix: address review findings - type consistency, thread safety, struct.error
- Change tcm_command_buf_length from uint8_t to uint16_t for consistency
  with other buffer length counters
- Replace bare bool self.running with threading.Event for proper
  cross-thread visibility (was relying on CPython GIL implementation detail)
- Add struct.error to exception handlers in thread loops (struct.error
  inherits from Exception, not OSError/ValueError, so malformed packets
  would kill the receiver thread)

+++ laser_engine/laser_engine.ino
@@ -165,7 +165,7 @@ uint8_t gProcessChannelStatusIndex = 0;
 bool reply_frame_analyzing_flag = false;
 
 char tcm_command_buf[256];
-uint8_t tcm_command_buf_length = 0;
+uint16_t tcm_command_buf_length = 0;
 
 char tcm_reply_buf[256];
 uint16_t tcm_reply_buf_length = 0;
